moved Tools scripts folder
