+++ Assets/_Application/Configuration/Settings.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3576bcc32ad0b4240969e5c3977c9f17
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
