+++ Assets/_Application/IO.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a9bbce4bd45894166b79aafcd6f64b1f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
