+++ Assets/_Application/Tools/FunctionGroup.cs
@@ -1,10 +1,6 @@
-using UnityEditor;
 using UnityEngine;
-using UnityEngine.Events;
-using UnityEngine.Serialization;
-using UnityEngine.UIElements;
 
-namespace Netherlands3D.Twin
+namespace Netherlands3D.Twin.Tools
 {
     [CreateAssetMenu(menuName = "Netherlands3D/FunctionGroup", fileName = "FunctionGroup", order = 1)]
     public class FunctionGroup : ScriptableObject{}
