+++ Assets/_Application/Tools/UI/Inspector.cs
@@ -1,9 +1,7 @@
-using System.Collections;
-using System.Collections.Generic;
 using TMPro;
 using UnityEngine;
 
-namespace Netherlands3D.Twin
+namespace Netherlands3D.Twin.Tools.UI
 {
     public class Inspector : MonoBehaviour
     {
