+++ Assets/_Application/Tools/UI/OverlayInstantiator.cs
@@ -1,9 +1,8 @@
 using Netherlands3D.Twin.Layers;
-using Netherlands3D.Twin.Layers.LayerTypes;
 using UnityEngine;
 using UnityEngine.Serialization;
 
-namespace Netherlands3D.Twin.UI.LayerInspector
+namespace Netherlands3D.Twin.Tools.UI
 {
     public class OverlayInstantiator : MonoBehaviour
     {
