+++ Assets/_Functionalities/AreaDownload/Scripts/AreaSelection.cs
@@ -6,7 +6,6 @@ using System.Runtime.InteropServices;
 using System.Text;
 using Netherlands3D.Collada;
 using Netherlands3D.MeshClipping;
-using Netherlands3D.Twin;
 using UnityEngine;
 using UnityEngine.Events;
